Fix links to search based on tags in README.md
The lables in the README.md seem to have changed and did no longer
work. Therefore the search queries where adjusted.

Additionally the search for the tags "Single Mirror" and "Offline
Mirror" was split into to seperate links as GitHub uses the AND
operator instead of OR (which is not supported) which would result in
wrong search results.